+++ packages/giochi/src/index.ts
@@ -212,6 +212,6 @@ export {
   type IdGiocoSala,
   type Listino,
 } from "./borsa";
-export { entra, evento, giocaCarta, pescaCarta, segnaPunti, stacca, statoSala, type StatoSala } from "./sala";
+export { azzeraPartita, entra, evento, giocaCarta, pescaCarta, ricarica, segnaPunti, stacca, statoSala, type StatoSala } from "./sala";
 export type { Borsa, Giudizio, OraDiBorsa, Partita, Stacco } from "./tipi";
 export { fileDellaSala } from "./sala-file";